% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_system.R
\name{abort}
\alias{abort}
\title{Throws an error}
\usage{
abort(..., class = NULL, call = FALSE)
}
\arguments{
\item{...}{The message. This can be a character string or a
combination of character strings and variables.}

\item{class}{An optional class or vector of classes to add to the error.}

\item{call}{If TRUE, the call will be included in the error message. Default
is FALSE.}
}
\description{
This is a convenience wrapper for internal use.
}
\keyword{internal}
