% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-add_dic_class.R
\name{add_dic_class}
\alias{add_dic_class}
\title{Add dictionary class to a vector or data.frame}
\usage{
add_dic_class(data)
}
\arguments{
\item{data}{A vector or a data.frame (or an object that inherits from a
data.frame) containing dic attributes.}
}
\value{
An object inheriting from class "dic". If a data.frame is provided,
a data.frame with columns inheriting from class "dic".
}
\description{
Adds the class 'dic' to a vector or all variables within a vector if it
contains 'dic' attributes but does not yet inherit from class 'dic'.
This is useful when dic attributes have been added to a vector or data.frame
but the class 'dic' has been removed (e.g., by subsetting).
}
\details{
If a data.frame is provided, all columns are checked for dic attributes and
the class 'dic' is added to those columns that have dic attributes but do not
yet inherit from class 'dic'.
}
\examples{
item <- remove_dic_class(ex_itrf$itrf_I_1)
item
item |> add_dic_class()
}
\keyword{internal}
