% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_system.R
\name{add_message}
\alias{add_message}
\title{Add a message to the message system}
\usage{
add_message(..., collapse = "", frame = -1, detail = 1, warning = FALSE)
}
\arguments{
\item{collapse}{A string to separate the components of the message.}

\item{frame}{An integer or character indicating the frame from which the message originates.}

\item{detail}{An integer indicating the detail level of the message.}

\item{warning}{Logical. If TRUE, the message is treated as a warning.}
}
\value{
None. The function modifies the internal message environment.
}
\description{
This function adds a message to the internal message system. Messages can be printed later using \code{print_messages()}.
#' @param ... Components of the message to be concatenated.
}
\examples{
init_messages()
add_message("This is a test message.")
print_messages()
}
\keyword{internal}
