% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_dic.R
\name{apply_dic}
\alias{apply_dic}
\title{Apply dictionary to data frame}
\usage{
apply_dic(
  data,
  dic,
  set_label_attr = TRUE,
  coerce_class = TRUE,
  replace_missing = TRUE,
  score_scales = FALSE,
  check_values = FALSE,
  impute_values = FALSE,
  rename_var = NULL,
  format_date = "\%Y-\%m-\%d"
)
}
\arguments{
\item{data}{Data frame or a character string with a filename (for now an
Microsoft Excel file) containing the data.}

\item{dic}{A data frame comprising a dictionary or a character string with a
filename (for now an Microsoft Excel file) containing a dictionary.}

\item{set_label_attr}{If TRUE, label attributes from the haven package are
set. These labels are shown in the Rstudio View panel.}

\item{coerce_class}{If set TRUE mismatches between class and dic type are
corrected.}

\item{replace_missing}{If TRUE, missing values from the dic are replaced with
NAs in the data frame.}

\item{score_scales}{Deprecated. Use \code{score_from_dic()} instead.}

\item{check_values}{If TRUE, performs the check_values function on the
variables of the data frame included in the dic file.}

\item{impute_values}{If TRUE and score_scales is TRUE, missing values are
automatically imputed based on scale information provided in the dic file.}

\item{rename_var}{When a character is provided, corresponding column from the
dic-file is used to rename variables from 'rename_var' (old) to 'item_name'
(new) in the \code{data} data frame.}

\item{format_date}{Optional string that is applied when character variable is
coerced to a 'Date' class.}
}
\value{
A data frame with dictionary information.
}
\description{
Joins a data frame with a dictionary file. The dictionary file provides
information about each variable in the data frame such as variable labels,
value labels, missing values, and scale definitions. The function applies
this information to the data frame. Optionally, missing values are replaced
with NAs, and scales are scored.
}
\details{
This is a key function in the \code{scaledic} package as it prepares a data
frame for further analyses by applying the dictionary information.
}
\examples{
dat <- apply_dic(dat_itrf, dic_itrf)
descriptives(dat)
}
\seealso{
\code{\link{new_dic}}, \code{\link{replace_missing}},
\code{\link{score_from_dic}}, \code{\link{check_values
 }}, \code{\link{dic_haven}}
}
\author{
Juergen Wilbert
}
