% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_dic.R
\name{as_dic}
\alias{as_dic}
\title{Convert a data frame to a dictionary-like structure}
\usage{
as_dic(x)
}
\arguments{
\item{x}{A data frame to be converted.}
}
\value{
A data frame with each column converted to a 'dic' object.
}
\description{
Converts each column of the input data frame into a 'dic' object, preserving
any existing labels as item labels.
}
