% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_deploy.R
\name{backup_dic}
\alias{backup_dic}
\title{Backup dictionary information}
\usage{
backup_dic(data)
}
\arguments{
\item{data}{A data frame containing dic infromation.}
}
\value{
A list with dic information. This can later be writen back to a data frame with the deploy_dic() function.
}
\description{
Backup dictionary information
}
