% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-build_scaledic_skeleton.R
\name{build_scaledic_skeleton}
\alias{build_scaledic_skeleton}
\title{Build a dictionary file template for scale definition}
\usage{
build_scaledic_skeleton(filename = NA, nrows = 0)
}
\arguments{
\item{filename}{Character string. Default is 'dic_template.xlsx'. If NA, returns
a data frame instead of writing an Excel file.}

\item{nrows}{Number of empty rows added to data frame. Default is 0.}
}
\value{
When 'filename' is not empty, it writes an Excel file with an empty
template of a dic file. When 'filename' is NA, returns a data
frame.
}
\description{
This function builds an empty skeleton of a dictionary file for scale
definition. It can either write an Excel file or return a data frame.
}
\details{
The skeleton contains all necessary variable names for a dic file.
The user can specify the number of empty rows to be added to the data frame.
This is useful when creating a dic file for a data frame with many variables.
The function uses the variable names defined in the option 'dic_file_vars'.
These can be modified by the user if needed.
The function uses the 'openxlsx' package to write the Excel file.
Make sure to have it installed.
}
\examples{
build_scaledic_skeleton(NA, nrows = 3)
}
\keyword{internal}
