% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_values.R
\name{check_values}
\alias{check_values}
\title{Check values}
\usage{
check_values(
  data,
  replace = NULL,
  return = TRUE,
  report = TRUE,
  include_missing = FALSE,
  integer_as_float = FALSE,
  check_type = TRUE
)
}
\arguments{
\item{data}{A data frame.}

\item{replace}{Value which replaces invalid values (e.g., NA).}

\item{return}{If TRUE, a data frame is returned with replaced values.}

\item{report}{If TRUE, an overview of invalid values will be given.}

\item{include_missing}{If TRUE, missing values (provided as 'missing' in the
dic file) will be considered as valid values.}

\item{integer_as_float}{If TRUE, type 'integer' will be handled as 'float'.
That is, only values outside the minimum and the maximum of the provided
valid values will be considered invalid.}
}
\value{
A data frame with replaced values if \code{replaces} is not NULL.
}
\description{
Checks if values in variables are valid according to the 'values' and 'type'
dictionary attributes. Invalid values can be replaced with a specified value.
}
\examples{
check_values(ex_itrf, return = FALSE)
}
