% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dic.R
\name{create_dic_attributes}
\alias{create_dic_attributes}
\title{Build and validate dic attribute list}
\usage{
create_dic_attributes(
  x = NULL,
  item_name = NULL,
  item_label = NULL,
  values = NULL,
  value_labels = NULL,
  missing = NULL,
  weight = 1,
  type = NULL,
  recodes = NULL,
  class = "item",
  ...
)
}
\arguments{
\item{x}{Optional. Used for type estimation only if \code{type} missing.}
}
\description{
Build and validate dic attribute list
}
\keyword{internal}
