% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaledic_data.R
\docType{data}
\name{dat_itrf}
\alias{dat_itrf}
\title{Example dataset: Integrated Teacher Report Form}
\format{
A data frame with 4776 rows and 40 variables

A data frame with 4776 rows and 40 variables
}
\usage{
dat_itrf

dat_itrf
}
\description{
A dataset including dic information containing the response teacher form data (40 items).

A dataset without dic information containing the response teacher form data (40 items).
}
\keyword{datasets}
