% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-descriptives.R
\name{descriptives}
\alias{descriptives}
\title{Table with descriptive statistics}
\usage{
descriptives(data, round = 2, labels = FALSE)
}
\arguments{
\item{data}{A data frame}

\item{round}{Digits for round function}

\item{labels}{Deprecated. Use \code{\link[=rename_items]{rename_items()}} instead.}
}
\value{
A data frame with descriptive statistics
}
\description{
Computes descriptive statistics for all numeric variables in a data frame.
}
\examples{
descriptives(ex_itrf)
}
