% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dic_attr.R
\name{dic_attr}
\alias{dic_attr}
\alias{dic_attr<-}
\title{Set and get dictionary attributes}
\usage{
dic_attr(x, var)

dic_attr(x, var) <- value
}
\arguments{
\item{x}{A vector or a data.frame.}

\item{var}{Attribute name. If missing, the full attribute list is returned.}

\item{value}{set value.}
}
\value{
Attribute value

Modified object with updated attribute.
}
\description{
Get or set dictionary attributes of a vector or a data.frame.
}
\examples{
# set dic attributes
x <- 1:5
dic_attr(x, "class") <- "item"
dic_attr(x, "item_label") <- "An example item"
dic_attr(x, "item_label")
dic_attr(x)
}
