% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haven.R
\name{dic_haven}
\alias{dic_haven}
\title{Create haven labels and value labels from dic}
\usage{
dic_haven(data, overwrite = TRUE)
}
\arguments{
\item{data}{A data frame containing \code{dic} information}

\item{overwrite}{Logical. If \code{TRUE}, overwrites existing haven labels.}
}
\value{
A data frame with haven labels and value labels
}
\description{
Create haven labels and value labels from dic
}
\examples{
ex_itrf_copy <- haven_dic(ex_itrf, remove_haven_class = TRUE)
ex_itrf_copy <- dic_haven(ex_itrf)
identical(ex_itrf, ex_itrf_copy)
}
