% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITRF.R
\docType{data}
\name{dic_ITRF}
\alias{dic_ITRF}
\title{Dictionary file for the Integrated Teacher Report Formula.}
\format{A data frame with 39 rows and 16 variables}
\usage{
dic_ITRF
}
\description{
A dictionary file of the German version of the ntegrated Teacher Report Formula (ITRF).
}
\keyword{datasets}
