% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dic.R
\name{.expand_colon}
\alias{.expand_colon}
\title{Expand colon notation in values
E.g. "1, 3:5, 7" -> c(1, 3, 4, 5, 7)}
\usage{
.expand_colon(values, msg_prefix, type)
}
\description{
Expand colon notation in values
E.g. "1, 3:5, 7" -> c(1, 3, 4, 5, 7)
}
\keyword{internal}
