% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dic.R
\name{.extract_scores}
\alias{.extract_scores}
\title{Parse a \code{recodes} specification from a dic file
Supported syntax:
\itemize{
\item "1 = 5, 2 = 4, 3 = 3, 4 = 2, 5 = 1"
\item "1 = 0, 2 = 0, 3 = 1, 4 = 1, .default = 9"
No code evaluation is performed.
}}
\usage{
.extract_scores(recodes)
}
\description{
Parse a \code{recodes} specification from a dic file
Supported syntax:
\itemize{
\item "1 = 5, 2 = 4, 3 = 3, 4 = 2, 5 = 1"
\item "1 = 0, 2 = 0, 3 = 1, 4 = 1, .default = 9"
No code evaluation is performed.
}
}
\keyword{internal}
