% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dic-parser.R
\name{.extract_value_labels}
\alias{.extract_value_labels}
\title{Parse a \code{value_labels} specification from a dic file
Supported syntax
\itemize{
\item "1 = 'Yes'; 2 = 'No'; 3 = 'Maybe'"
\item "1 = Yes; 2 = No; 3 = Maybe"
\item "
1 = 'Yes'
2 = 'No'
3 = 'Maybe'
"
No code evaluation is performed.
}}
\usage{
.extract_value_labels(value_labels, type)
}
\description{
Parse a \code{value_labels} specification from a dic file
Supported syntax
\itemize{
\item "1 = 'Yes'; 2 = 'No'; 3 = 'Maybe'"
\item "1 = Yes; 2 = No; 3 = Maybe"
\item "
1 = 'Yes'
2 = 'No'
3 = 'Maybe'
"
No code evaluation is performed.
}
}
\keyword{internal}
