% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dic-parser.R
\name{.extract_values}
\alias{.extract_values}
\title{Parse a \code{values} specification from a dic file}
\usage{
.extract_values(values, type, item, field)
}
\description{
Supported syntax:
\itemize{
\item Integers: "1, 2, 3"
\item Integer ranges: "5:11" (inclusive)
\item Mixed: "1, 3:5, 9"
\item Float: "min, max" (two numeric values)
\item Character/factor: "'m', 'f', 'd'" or "\"m\", \"f\""
}
}
\details{
No code evaluation is performed.
}
\keyword{internal}
