% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaledic_data.R
\docType{data}
\name{ex_itrf}
\alias{ex_itrf}
\title{Example dataset: Integrated Teacher Report Form}
\format{
A data frame with 4776 rows and 40 variables
}
\usage{
ex_itrf
}
\description{
A dataset including dic information containing the response teacher form
data (40 items).
}
\keyword{datasets}
