% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaledic_data.R
\docType{data}
\name{ex_normtable_ext}
\alias{ex_normtable_ext}
\title{Example dataset with a norm table}
\format{
A data frame with 52 rows and 5 variables
}
\usage{
ex_normtable_ext
}
\description{
Example dataset with a norm table
}
\keyword{datasets}
