% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaledic_data.R
\docType{data}
\name{ex_normtable_int}
\alias{ex_normtable_int}
\title{Example of a norm table}
\format{
A data frame with 48 rows and 5 variables
}
\usage{
ex_normtable_int
}
\description{
Normatable for internalising scale
}
\keyword{datasets}
