% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaledic_data.R
\docType{data}
\name{ex_scaledic_data}
\alias{ex_scaledic_data}
\alias{ex_scaledic_dic}
\title{Example dataset and dic file: Religious beliefs and suicide tendency}
\format{
A data frame with 20 rows and 12 variables
}
\usage{
ex_scaledic_data
}
\description{
A dataset without dic information containing 20 students (12 items).
}
\keyword{datasets}
