% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_dic.R
\name{extract_dic}
\alias{extract_dic}
\title{Extract a dictionary from a data file with dic information}
\usage{
extract_dic(data)
}
\arguments{
\item{data}{A data frame with dic information.}
}
\value{
A data frame in a dictionary format.
}
\description{
This function extracts the dictionary information from a data frame
containing dic attributes and returns it as a data frame in a dictionary
format.
}
\examples{
## extract dic from ex_itrf
ex_itrf_dic <- extract_dic(ex_itrf)
head(ex_itrf_dic)
}
