% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dic_attribute.R
\name{get_dic_attribute}
\alias{get_dic_attribute}
\title{Get a dictionary attribute of one or more variables.}
\usage{
get_dic_attribute(data, attribute, duplicates = TRUE)
}
\arguments{
\item{data}{A data frame with dic information.}

\item{attribute}{Name of label attribute to retrieve.}

\item{duplicates}{If FALSE, duplicated attributes will be removed from
returned vector.}
}
\value{
Vector of attributes. If an attribute is not present, NA is returned.
}
\description{
Retrieves a specific dic attribute from all variables in a data frame.
}
\details{
If an attribute is not present for a variable, NA is returned for that variable.
If duplicates is set to FALSE, only unique attribute values are returned.
}
\examples{
ex_itrf |>
  select_items(subscale == "Int") |>
  get_dic_attribute("item_label")
}
