% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_index.R
\name{get_index}
\alias{get_index}
\title{Get index}
\usage{
get_index(data, names = TRUE, ...)
}
\arguments{
\item{data}{A data frame}

\item{names}{If names is TRUE, a vector with variable names will be returned.}

\item{...}{Any dic attribute name with a string vector (e.g., scale = "ITRF", subscale = c("Ext", "Int"), subscale = "ADB")}
}
\value{
A vector with indices adhering to the given attribute criteria.
}
\description{
Get index
}
