% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private-get_index_from_dic.R
\name{get_index_from_dic}
\alias{get_index_from_dic}
\title{Get index of items from dic based on filter expression}
\usage{
get_index_from_dic(data, filter, names = TRUE, class = NULL)
}
\arguments{
\item{data}{A data frame}

\item{filter}{A character string or expression to filter the dic}

\item{names}{If TRUE, variable names are returned instead of indices}

\item{class}{If provided, class will be added to the filter expression}
}
\value{
A vector with indices or names of variables adhering to the filter
criteria.
}
\description{
Get index of items from dic based on filter expression
}
\keyword{internal}
