% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_label.R
\name{get_label}
\alias{get_label}
\title{Get label of a scale}
\usage{
get_label(data, scale, label)
}
\arguments{
\item{data}{A data frame with dic information}

\item{scale}{A scale definition (e.g. scale = subscale == "Int").}

\item{label}{Name of label attribute}
}
\value{
Labels for a given scale
}
\description{
Get label of a scale
}
