% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scales.R
\name{get_scales}
\alias{get_scales}
\title{Extracts a list of item names based on dic information}
\usage{
get_scales(data, ...)
}
\arguments{
\item{data}{A data.frame with dic information.}

\item{...}{A logical expression defining a selection definition or a name
defining a dic attribute for automatic scale definition.}
}
\value{
A (named) list with character vectors of item names.
}
\description{
Extracts a list of item names based on dic information
}
\details{
This function is basically a wrapper around the \code{select_items(data = data, filter = ..., names_only = TRUE)} function. It takes multiple filter
expressions for a single data frame and returns the item names. This is
mostly helpful for functions that take multiple scale definitions like
\code{alpha_table()}. If you provide a name instead of a logical expression,
that name must be a dic attribute. A list of scales will be created based
on all values of that attribute.
}
\examples{
## define individual scales
get_scales(ex_itrf,
  'APD' = subscale_2 == "APD",
  'OPP' = subscale_2 == "OPP",
  "SW" = subscale_2 == "SW",
  'AD' = subscale_2 == "AD"
)

## generate scale list based on all instances of an attribute
get_scales(ex_itrf, subscale_2)
}
