% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haven.R
\name{haven_dic}
\alias{haven_dic}
\title{haven labels to dic files}
\usage{
haven_dic(data, remove_haven_class = FALSE)
}
\arguments{
\item{data}{A data frame containing variables with haven labels.}

\item{remove_haven_class}{If TRUE, haven labels are removed.}
}
\value{
A data frame with dic information.
}
\description{
Convert haven labelled data frames to dic files
}
\details{
This function converts a data frame with haven labels into a data frame
containing dic information. Item labels and value labels are extracted from
the haven attributes and stored in dic attributes.
}
\examples{
ex_itrf_dic <- haven_dic(ex_itrf, remove_haven_class = TRUE)
ex_itrf_copy <- dic_haven(ex_itrf_dic)

}
