% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_values.R
\name{id_invalid_values}
\alias{id_invalid_values}
\title{Identify invalid values in a vector according to dic attributes}
\usage{
id_invalid_values(x, include_missing = FALSE)
}
\arguments{
\item{x}{A dic vector}

\item{include_missing}{If TRUE, missing values (provided as 'missing' in the
dic file) will be considered as valid values. If FALSE, missing values will
be considered as invalid values.}
}
\value{
A vector with indices of invalid values
}
\description{
Identify invalid values in a vector according to dic attributes
}
\keyword{internal}
