% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_dic.R
\name{is_dic}
\alias{is_dic}
\title{Check if an object is of class 'dic'}
\usage{
is_dic(x)
}
\arguments{
\item{x}{An object to check.}
}
\value{
TRUE if the object is of class 'dic', FALSE otherwise.
}
\description{
Checks if the provided object inherits from the 'dic' class.
}
