% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaledic_data.R
\docType{data}
\name{ITRF}
\alias{ITRF}
\alias{ex_itrf}
\alias{dic_ITRF}
\title{Data ITRF}
\format{
An object of class \code{data.frame} with 4776 rows and 36 columns.
}
\usage{
ITRF
}
\description{
Data of the ITRF and a correponding dic file'
}
\keyword{datasets}
