% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_dic.R
\name{keep_dic}
\alias{keep_dic}
\title{Keep attributes of a dictionary when evaluating an expression}
\usage{
keep_dic(x, expression, envir = parent.frame(n = 1))
}
\arguments{
\item{x}{A dictionary object.}

\item{expression}{An expression to evaluate. This can be a character string
or an expression}

\item{envir}{The environment in which to evaluate the expression. Defaults to
the parent frame.}
}
\value{
A dictionary object with the same attributes as the original, but
with the expression evaluated.
}
\description{
This function evaluates an expression while keeping the attributes of a
dictionary object intact. It is useful when you want to perform operations on
a dictionary without losing its attributes.
}
