% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_scales.R
\name{list_scales}
\alias{list_scales}
\title{List scales in a data frame based on dic attributes}
\usage{
list_scales(
  data,
  levels = c("scale", "subscale", "subscale_2"),
  n_items = FALSE,
  char_na = ""
)
}
\arguments{
\item{data}{The target data frame.}

\item{levels}{Character vector with names of dic attributes used to extract scale information.}

\item{n_items}{If TRUE, number of items for each scale is shown.}

\item{char_na}{Character for NAs.}
}
\value{
A data.frame with scales on different levels.
}
\description{
Lists scales and subscales defined in the dic attributes of a data frame.
This is useful for getting an overview of the scales present in a dataset
that uses dic attributes to define scales.
}
\details{
The function extracts scale information from dic attributes of variables
in the data frame. By default, it looks for attributes named "scale",
"subscale", and "subscale_2". It can also count the number
of items for each scale or subscale if requested.
}
