% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_norms.R
\name{lookup_norms}
\alias{lookup_norms}
\title{Look up norm table values based on raw scores and group affiliations}
\usage{
lookup_norms(
  rawscores,
  group = NULL,
  normtable,
  from = "raw",
  to = "T",
  group_label = names(group),
  label = NULL
)
}
\arguments{
\item{rawscores}{A vector with raw scores.}

\item{group}{A vector with group affiliations or a list with vectors for
multiple group categorizations. If NULL (default), no group
affiliations are used.}

\item{normtable}{An excel file name or a data frame containing a norm table.}

\item{from}{Label of the raw score variable in file.}

\item{to}{Label of the norm score variable in file.}

\item{group_label}{Label of the group variable in file or a list with group
labels for multiple group categorizations. If NULL (default), "group" is used
as group label.}

\item{label}{Item label of the resulting variable. If NULL (default), no label is set.}
}
\value{
A vector with norm values. If ambiguities occur, NA is returned
for the respective raw score(s) and a message is printed.
}
\description{
Transforms raw values to norm values based on a norm table.
If group affiliations are provided, these are used to identify the correct
norm values. If no group affiliations are provided, norm values are looked up
based on raw scores only. In case of ambiguities (e.g., multiple norm
values for the same raw score), NA is returned and a message is printed.
}
\examples{
normtable <- data.frame(
  age = rep(c(6, 8, 6, 8), each = 11),
  gender = rep(c("m", "w"), each = 22),
  raw =  rep(0:10, 4),
  T = rep(c(40, 42, 44, 46, 48, 50, 52, 54, 56, 58, 60,
            37, 39, 41, 43, 45, 47, 49, 51, 53, 55, 57), 2) + rep(c(0,5), each = 22)
)
rawscores <- c(5,5,3,1)
group_age <- c("6", "8", "6", "8")
group_gender <- c("m", "m", "w", "w")

lookup_norms(rawscores, group = list(age = group_age, gender = group_gender), normtable)

## When group values are not specified exactly, raw scores can be ambiguous:
lookup_norms(rawscores, group = list(gender = group_gender), normtable = normtable)

lookup_norms(rawscores, normtable = normtable)
}
