% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_dic.R
\name{new_dic}
\alias{new_dic}
\title{Low-level constructor for dic vectors}
\usage{
new_dic(x, dic_attributes = list(), label = NULL, labels = NULL)
}
\arguments{
\item{x}{An atomic vector (integer/double/logical/character/factor/Date, ...)}

\item{dic_attributes}{A named list with dic metadata.}

\item{label}{Optional haven-style variable label.}

\item{labels}{Optional haven-style value labels (named vector).}
}
\value{
A dic vector with specified attributes.
}
\description{
Creates a dic vector with specified attributes.
Detailed parsing and validation is not performed by this function.
Use the \code{dic()} factory function for that purpose.
}
\details{
Details:
\itemize{
\item The function sets the class of \code{x} to include "dic".
\item The provided \code{dic_attributes} list is assigned as the "dic" attribute.
\item Optional haven-style variable label and value labels can be set.
}
}
\keyword{internal}
