% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_system.R
\name{notify}
\alias{notify}
\title{Throws a message}
\usage{
notify(..., type = "!", detail = 1, warning = FALSE)
}
\arguments{
\item{...}{The message. This can be a character string or a
combination of character strings and variables.}

\item{detail}{The level of detail for the message. Messages with a detail
level higher than the current detail level set in the options will not be
shown. The default is 1.}

\item{warning}{If TRUE, the message will be shown as a warning. If FALSE
(default), the message will be shown as an informational message.}
}
\description{
Messages can be shown with different levels
of detail and as warnings. This is a convenience wrapper for internal use.
}
\keyword{internal}
