% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dic.R
\name{print.dic}
\alias{print.dic}
\title{Print dic infos}
\usage{
\method{print}{dic}(x, ...)
}
\arguments{
\item{x}{A variable with dic infos}

\item{...}{Further parameters for the print function}
}
\value{
Dic infos of x
}
\description{
Print dic infos
}
\keyword{internal}
