% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dic.R
\name{print_dic}
\alias{print_dic}
\title{Prints dic infos of a variable with dic infos.}
\usage{
print_dic(x, length = 100, ...)
}
\arguments{
\item{x}{A variable with dic infos.}

\item{length}{Number of entries to be printed.}

\item{...}{Unused future parameters.}
}
\value{
Dic infos of x and the first \code{length} entries of \code{x}.
}
\description{
Prints dic infos of a variable with dic infos.
}
