% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych_fa.R
\name{psych_fa}
\alias{psych_fa}
\title{Factor analyses based on the psych::fa function}
\usage{
psych_fa(..., sort = TRUE, cut = 0.2, round = 2, type = "df")
}
\arguments{
\item{...}{Arguments passed to the psych::fa function}

\item{sort}{If TRUE, loadings are sorted}

\item{cut}{loadings below cut will be ommited in the resulting data frame}

\item{round}{number of digitis to round loadings (based on the base::round function)}

\item{type}{If 'html' returns an html table. Otherwise a data.frame}
}
\value{
A data.frame
}
\description{
Factor analyses based on the psych::fa function
}
