% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private.R
\name{read_by_suffix}
\alias{read_by_suffix}
\title{Decides based on suffix which function to use for reading a data.file}
\usage{
read_by_suffix(filename)
}
\description{
Decides based on suffix which function to use for reading a data.file
}
\keyword{internal}
