% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_dic_class.R
\name{remove_dic_class}
\alias{remove_dic_class}
\title{Remove dictionary class and information from a data.frame or a vector}
\usage{
remove_dic_class(data, remove_attributes = FALSE)
}
\arguments{
\item{data}{A vector or a data.frame (or an object that inherits from a data.frame)}

\item{remove_attributes}{If TRUE, all dictionary attributes are removed.}
}
\value{
An object not inheriting from class "dic".
}
\description{
Remove dictionary class and information from a data.frame or a vector
}
\examples{
dat <- remove_dic_class(ex_itrf)
class(ex_itrf$itrf_I_1)
class(dat$itrf_I_1)
}
