% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_by_list.R
\name{rename_by_list}
\alias{rename_by_list}
\title{Rename varibles based on a list}
\usage{
rename_by_list(data, file = NULL, to_from = NULL, to = NULL, from = NULL)
}
\arguments{
\item{data}{A data frame with variables to be renamed}

\item{file}{A filename of an Excel file containing at least two columns (variables) for renaming. Or a data frame with two columns.}

\item{to_from}{When a filename or a data.frame is provided, a named charcter with the names of the target  and source variable names (e.g., c("to" = "from")).
When no filename is provided, to_from must be a vector with named variable names c("to1" = "from1", "to2" = "from2")).}

\item{to}{When a filename or a data.frame is provided, the name of the column with the target variable names. When no filename is provided, to must a vector with target variable names.}

\item{from}{When a filename or a data.frame is provided, the name of the column with the source variable names. When no filename is provided, from must a vector with source variable names.}
}
\value{
A data frame
}
\description{
Rename varibles based on a list
}
\examples{
dat <- data.frame(A = NA, B = NA, C = NA, D = NA)
rename_by_list(dat, to = c("albert", "bea"), from = c("A", "B"))
rename_by_list(dat, to_from = c("carl" = "C", "daniel" = "D"))
dic <- data.frame(old = c("A", "B"), new = c("albert", "bea"))
rename_by_list(dat, dic, to_from = c("new" = "old"))
rename_by_list(dat, dic, to = "new", from = "old")
\dontrun{
rename_by_list(dat, "rename_list,xlsx", to_from = c("new" = "old"))
rename_by_list(dat, "rename_list,xlsx", to = "new", from = "old")
}
}
