% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_item.R
\name{rename_item}
\alias{rename_item}
\title{Rename items
Rename items based on dic information.}
\usage{
rename_item(
  data,
  pattern = "label",
  chars = NULL,
  char_sep = ": ",
  char_weight = c("(-)", "(+)"),
  char_prefix_end = ": "
)
}
\arguments{
\item{data}{A data frame}

\item{pattern}{A character string or vector of character strings defining a prefix.
May include "label", "scale", "subscale", "subscale_2", "reverse", "weight", or "name".}

\item{chars}{If not NULL, only the first n chars og the long label will be applied.}

\item{char_sep}{Character with seperator between prefix information.}

\item{char_weight}{Character vector of length two with signs for negative and positive weights.}
}
\value{
A renamed data frame
}
\description{
Rename items
Rename items based on dic information.
}
