% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_item_deprecated.R
\name{rename_items_deprecated}
\alias{rename_items_deprecated}
\title{Rename items (depracted version)}
\usage{
rename_items_deprecated(
  data,
  pattern = "item_label",
  chars = NULL,
  char_sep = "_",
  char_weight = c("(-)", "(+)"),
  char_prefix_label = ": "
)
}
\arguments{
\item{data}{A data frame}

\item{pattern}{A character string or vector of character strings defining a prefix.
May include the name of any dic attribute (e.g."item_label", "scale", "subscale", "subscale_2") or some shortcuts: "reverse", "label", or "name").}

\item{chars}{If not NULL, only the first n chars og the long label will be applied.}

\item{char_sep}{Character with seperator between prefix information.}

\item{char_weight}{Character vector of length two with signs for negative and positive weights.}
}
\value{
A renamed data frame
}
\description{
Rename items based on dic information.
}
\examples{
ex_itrf  |>
  rename_items(pattern = c("reverse", "label")) |>
  names()

}
