% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_missing.R
\name{replace_missing}
\alias{replace_missing}
\title{Replace Missing values}
\usage{
replace_missing(data, replace = NA, report = TRUE)
}
\arguments{
\item{data}{A data frame}

\item{replace}{Replace value for missing values (default is NA).}

\item{report}{If TRUE, proportion of missing values is printed.}
}
\value{
A data frame with replaced missing values.
}
\description{
This function replaces missing values in a data frame based on the
missing value definitions provided in the dic attributes of the variables.
Variables without dic attributes are ignored.
}
