% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_scaledic.R
\docType{package}
\name{scaledic-package}
\alias{scaledic-package}
\alias{scaledic}
\title{Scale Dictionary}
\description{
A collection of procedures to bring labels, scales, missing values etc. to a data frame.
}
\author{
Juergen Wilbert [aut, cre]
}
\keyword{package}
