% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_scores.R
\name{select_scores}
\alias{select_scores}
\title{Select scale scores
Selects a data frame with score variables generated by score_scale.}
\usage{
select_scores(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A data frame
}
\description{
Select scale scores
Selects a data frame with score variables generated by score_scale.
}
