% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_var.R
\name{select_var}
\alias{select_var}
\title{Subset a dic variable}
\usage{
select_var(x, i = length(x), ...)
}
\arguments{
\item{x}{An  object}

\item{i}{Elements to be selected}

\item{...}{further arguments passed to the subset function.}
}
\value{
A dic variable
}
\description{
Subset a dic variable
}
\keyword{internal}
