% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_label.R
\name{set_label}
\alias{set_label}
\title{Set dictionary information to variable}
\usage{
set_label(data, vars, parameter, values)
}
\arguments{
\item{data}{A data frame}

\item{vars}{string or vector with string of variable names}

\item{values}{Vector of values of the same length as vars}

\item{paramter}{Character string with dic parameter}
}
\value{
A data frame with dic information
}
\description{
Set dictionary information to variable
}
