% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-subset_dic.R
\name{[.dic}
\alias{[.dic}
\alias{[.dic<-}
\title{Subset a dic variable while preserving dic attributes}
\usage{
\method{[}{dic}(x, i = TRUE, ...)

\method{[}{`dic<-`}(x, i, ..., value)
}
\arguments{
\item{x}{An  object of class dic.}

\item{i}{Elements to be selected.}

\item{...}{further arguments passed to the subset function.}
}
\value{
A dic variable
}
\description{
This function allows subsetting of dic variables while preserving their dic attributes,
as well as any haven labels and value labels.
}
\keyword{internal}
