% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dic_class.R
\name{add_dic_class}
\alias{add_dic_class}
\title{Add dictionary class to a vector}
\usage{
add_dic_class(data)
}
\arguments{
\item{data}{A vector or a data.frame (or an object that inherits from a data.frame)}
}
\value{
An object inheriting from class "dic".
}
\description{
Adds the class 'dic' to a vector or all variables within a vector if it contains 'dic' attributes.
}
\examples{
item <- remove_dic_class(ex_itrf$itrf_I_1)
item
item |> add_dic_class()
}
