% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dic_class.R
\name{add_dic_class}
\alias{add_dic_class}
\title{Add dictionary class to a data.frame or a vector}
\usage{
add_dic_class(data)
}
\arguments{
\item{data}{A vector or a data.frame (or an object that inherits from a data.frame)}
}
\value{
An object not inheriting from class "dic".
}
\description{
Add dictionary class to a data.frame or a vector
}
