% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_table.R
\name{alpha_table}
\alias{alpha_table}
\title{Table with alpha values}
\usage{
alpha_table(
  data,
  scales,
  labels = NULL,
  round = 2,
  CI = TRUE,
  conf_level = 0.95,
  check_key = TRUE,
  omega = FALSE,
  keys = NULL,
  RMSEA = FALSE,
  difficulty = FALSE,
  values = NULL
)
}
\arguments{
\item{data}{A data Frame}

\item{scales}{A list containing vectors with variable names. Each list element defines one scale. Named list elements are used as labels.}

\item{labels}{Label names for scales (defaults to named list elements in 'scales').}

\item{round}{Rounds values to given decimal position.}

\item{CI}{If TRUE confidence intervals are calculated.}

\item{conf_level}{Confidence level (e.g. 0.95 for 95 percent).}

\item{check_key}{Check_key for the psych::alpha function.}

\item{omega}{If TRUE Omega reliability estimation is calculated.}

\item{keys}{Optional key argument for the psych::alpha function.}

\item{RMSEA}{If TRUE RMSEA is calculated.}

\item{values}{Sets maximum and minimum valid values necessary to calculate item difficulty}
}
\description{
Returns a list of alpha cronbachs.
}
