% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_dic.R
\name{apply_dic}
\alias{apply_dic}
\title{Apply dictionary}
\usage{
apply_dic(data, dic, factors = TRUE)
}
\arguments{
\item{data}{Data frame}

\item{dic}{Dictionary file}

\item{factors}{If set TRUE, factor variables will be turned into factors.}
}
\value{
A data frame with dictionary information.
}
\description{
Apply dictionary
}
\examples{
dat <- apply_dic(ITRF, dic_ITRF)
list_scales(dat)
}
