% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_dic.R
\name{apply_dic}
\alias{apply_dic}
\title{Apply dictionary}
\usage{
apply_dic(
  data,
  dic,
  factors = TRUE,
  set_label_attr = TRUE,
  coerce_class = TRUE,
  replace_missing = TRUE,
  score_scales = TRUE,
  check_values = FALSE,
  impute_values = FALSE,
  rename_var = NULL
)
}
\arguments{
\item{data}{Data frame}

\item{dic}{A data frame comprising a dictionary or a character string with a
filename (for now an Microsoft Excel file) containg a dictionary.}

\item{factors}{If set TRUE, factor variables will be turned into factors.}

\item{set_label_attr}{If TRUE, a separate label attribute (see the haven package) is set.}

\item{coerce_class}{If set TRUE mismatches between class and dic type are corrected.}

\item{replace_missing}{If TRUE, missing values from the dic are replaced with NA}

\item{score_scales}{If TRUE and the dic files contains score scale
definitions these are applied}

\item{check_values}{If TRUE, performs the check_values function on the variables of the data frame included in the dic file.}

\item{impute_values}{If TRUE and score scale informtion are available, missing values are imputed.}

\item{rename_var}{A character string providing a variable name. This variable contains variable names of the data files to be renames to the corresponding item_name from the dic file.}
}
\value{
A data frame with dictionary information.
}
\description{
Apply dictionary
}
\examples{
dat <- apply_dic(dat_itrf, dic_itrf)
descriptives(dat)
}
