% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_scaledic_skeleton.R
\name{build_scaledic_skeleton}
\alias{build_scaledic_skeleton}
\title{Build a dictionary file template}
\usage{
build_scaledic_skeleton(filename = "dic_template.xlsx", nrows = 0)
}
\arguments{
\item{filename}{Character string. Default is 'dic_template.xlsx'}

\item{nrows}{Number of empty rows added to data frame.}
}
\value{
When 'filename' is not empty, it writes an Excel file with an empty
template of a dic file. When 'filename' is NA, returns a data
frame.
}
\description{
Build a dictionary file template
}
\examples{
build_scaledic_skeleton(NA, nrows = 3)
}
