% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_scaledic_skeleton.R
\name{build_scaledic_skeleton}
\alias{build_scaledic_skeleton}
\title{Build a dictionary file template}
\usage{
build_scaledic_skeleton(filename = "dic_template.xlsx")
}
\arguments{
\item{filename}{Character string. Default is 'dic_template.xlsx'}
}
\value{
Writes an Excel file with an empty template of a dic file.
}
\description{
Build a dictionary file template
}
