% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_values.R
\name{check_values}
\alias{check_values}
\title{Check values
Checks if values in variables are within the set difined in
the values and type dictionary attributes.}
\usage{
check_values(data, replace = NULL, return = TRUE, report = FALSE,
  include_missing = TRUE)
}
\arguments{
\item{data}{A data frame}

\item{replace}{Value which relaces unvalid values (e.g., NA).}

\item{return}{If TRUE, a data frame is returned with replaced values.}

\item{report}{If TRUE, an overview of invalid values will be given.}

\item{include_missing}{If TRUE, missing values (provided in the dic file)
will be considered as valid values.}
}
\value{
A data frame with replaced values if replaces is not NULL.
}
\description{
Check values
Checks if values in variables are within the set difined in
the values and type dictionary attributes.
}
\examples{
dat <- apply_dic(ITRF, dic_ITRF)
check_values(dat, return = FALSE)
}
