% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data_frame.R
\name{combine_data_frames}
\alias{combine_data_frames}
\title{Combine data frames
Comnine data frames and keep dic information.}
\usage{
combine_data_frames(...)
}
\arguments{
\item{...}{data frames with dic information}
}
\value{
A combined data frame with all rows and columns from the provided dataframes.
}
\description{
Combine data frames
Comnine data frames and keep dic information.
}
