% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data_frame.R
\name{combine_data_frames}
\alias{combine_data_frames}
\title{Combine data frames}
\usage{
combine_data_frames(...)
}
\arguments{
\item{...}{data frames with dic information}
}
\value{
A combined data frame with all rows and columns from the provided
data frames.
}
\description{
Combine data frames (rows and columns) and keep dic information.
}
\details{
This function is useful when you want to join several data frames
that contain dic information into one data frame. When a variable that is
entailed in both data frames has dic information, the resulting data frame
will keep the dic information of the data frame that is first listed as an
argument.
}
