% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_dic.R
\name{deploy_dic}
\alias{deploy_dic}
\title{Deploy dictionary information}
\usage{
deploy_dic(data, dic_list)
}
\arguments{
\item{data}{A data frame}

\item{dic_list}{A list created by the get_dic() function.}
}
\value{
A data frame with added dic information from dic_list
}
\description{
Deploy dictionary information
}
