% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{descriptives}
\alias{descriptives}
\title{Table with descriptive statistics}
\usage{
descriptives(data, round = 2, labels = FALSE)
}
\arguments{
\item{data}{A data frame}

\item{round}{Digits for round function}

\item{labels}{If TRUE, item_labels are included in the table.}
}
\value{
A data frame with descriptive statistics
}
\description{
Table with descriptive statistics
}
\examples{
descriptives(ex_itrf)
}
