% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dic-constructor.R
\name{dic}
\alias{dic}
\title{Create a dic vector (factory)}
\usage{
dic(
  x,
  item_name = NULL,
  item_label = NULL,
  values = NULL,
  value_labels = NULL,
  missing = NULL,
  weight = 1,
  type = NULL,
  recodes = NULL,
  class = "item",
  ...,
  dic_attributes = NULL,
  .coerce_class = TRUE,
  .format_date = "\%Y-\%m-\%d"
)
}
\arguments{
\item{x}{A vector.}

\item{item_name}{Character.}

\item{item_label}{Character.}

\item{values}{Numeric or character vector with values. The vector can be
named.}

\item{value_labels}{Character of the form \verb{value = label; value2 = label2}}

\item{missing}{Numeric or character vector with values.}

\item{weight}{numeric.}

\item{type}{defaults to data type of x.}

\item{recodes}{Recoding information e.g. \verb{4 = 1, .default = 0}.}

\item{class}{default is "item".}

\item{dic_attributes}{Optional pre-built attribute list (skips parsing if given).}

\item{.coerce_class}{If TRUE, coerce x to match \code{type}.}

\item{.format_date}{Date format for coercion from character -> Date.}
}
\description{
Create a dic vector (factory)
}
