% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dic.R
\name{dic}
\alias{dic}
\title{Show variable with dic information}
\usage{
dic(data)
}
\arguments{
\item{A}{variable}
}
\value{
Prints content of varable with dic information added.
}
\description{
Show variable with dic information
}
