% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dic.R
\name{dic}
\alias{dic}
\title{Dictionary class constructor}
\usage{
dic(x, class = "item", type = "integer", ...)
}
\arguments{
\item{x}{A variable.}

\item{class}{default is "item".}

\item{type}{defaults to "integer".}

\item{...}{further dic parameters.}
}
\value{
An item of class dic.
}
\description{
Dictionary class constructor
}
\examples{
x <- dic(1:100, item_label = "The label of this item")
}
