% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dic_attr.R
\name{dic_attr}
\alias{dic_attr}
\alias{dic_attr<-}
\title{Set and get dictionary attributes}
\usage{
dic_attr(x, var)

dic_attr(x, var) <- value
}
\arguments{
\item{x}{Variable}

\item{var}{Attribute}

\item{value}{Value}

\item{values}{set value}
}
\value{
Attribute value
}
\description{
Set and get dictionary attributes
}
