% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dic_haven.R
\name{dic_haven}
\alias{dic_haven}
\title{Create haven labels and value labels from dic}
\usage{
dic_haven(data)
}
\arguments{
\item{data}{A data frame containing dic information}
}
\value{
A data frame with haven labels and value labels
}
\description{
Create haven labels and value labels from dic
}
