% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaledic_data.R
\docType{data}
\name{dic_itrf}
\alias{dic_itrf}
\title{Example dictionary file for the Integrated Teacher Report Form.}
\format{
A data frame with 39 rows and 15 variables
}
\usage{
dic_itrf
}
\description{
A dictionary file of the German version of the integrated Teacher Report
Form (ITRF).
}
\keyword{datasets}
