% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dic.R
\name{.check_coerce_class}
\alias{.check_coerce_class}
\title{Check and coerce class of x to match type}
\usage{
.check_coerce_class(x, attrs, .format_date)
}
\description{
Check and coerce class of x to match type
}
\keyword{internal}
