% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_dic.R
\name{enrich_dic}
\alias{enrich_dic}
\title{Enriches a data frame with dic information}
\usage{
enrich_dic(dat)
}
\arguments{
\item{dat}{A data frame}
}
\description{
Adds dic information to all variables in a data frame that do not have a dic
attribute.
}
