% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dic_attribute.R
\name{get_dic_attribute}
\alias{get_dic_attribute}
\title{Get a dictionary attribute of one or more variables.}
\usage{
get_dic_attribute(data, attribute, duplicates = TRUE)
}
\arguments{
\item{data}{A data frame with dic information}

\item{duplicates}{If FALSE, duplicated attributes will be removed from
returned vector.}

\item{attributes}{Name of label attribute}
}
\value{
Vector of attributes. If an attribute is not present, NA is returned.
}
\description{
Get a dictionary attribute of one or more variables.
}
\examples{
select_items(ex_itrf, subscale == "Int") |> get_dic_attribute("item_label")
}
