% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_index.R
\name{get_index}
\alias{get_index}
\title{Get index}
\usage{
get_index(
  data,
  filter = NULL,
  scale = NULL,
  subscale = NULL,
  subscale_2 = NULL,
  names = TRUE,
  class = NULL
)
}
\arguments{
\item{data}{A data frame}

\item{filter}{A logical expression for any dic attribute (e.g. scale == "ITRF" & subscale == "Int")}

\item{names}{If names is TRUE, a vector with variable names will be returned.}
}
\value{
A vector with indices adhering to the given attribute criteria.
}
\description{
Get index
}
