% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_index.R
\name{get_index}
\alias{get_index}
\title{Get index}
\usage{
get_index(data, scale = NULL, subscale = NULL, subscale_2 = NULL,
  names = TRUE)
}
\arguments{
\item{data}{A data frame}

\item{scale}{String with scale name}

\item{subscale}{String with subscale name}

\item{subscale_2}{String with sub subscale name}

\item{names}{If names is TRUE, a vector with variable names will be returned.
Otherwise a vector with variable indices.}
}
\value{
A vector with indices
}
\description{
Get index
}
