% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scales.R
\name{get_scales}
\alias{get_scales}
\title{Extracts a list of item names based on dic information}
\usage{
get_scales(data, ...)
}
\arguments{
\item{data}{A data.frame with dic inforation.}

\item{...}{selection definitions.}
}
\value{
A (named) list with string vectors of item names
}
\description{
Extracts a list of item names based on dic information
}
\details{
This function is basically a wrapper around the `
select_items(data = data, filter = ..., names_only = TRUE)` function.
It takes mutiple filter expressions for a single data frame and returns the item names.
This is mostly helpfull for functions that take multiple scale definitions like `alpha_table()`
}
