% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scales.R
\name{get_scales}
\alias{get_scales}
\title{Extracts a list of item names based on dic information}
\usage{
get_scales(data, ..., .variable = "item_name")
}
\arguments{
\item{data}{A data.frame with dic information.}

\item{...}{selection definitions.}

\item{.variable}{character with name of the dic information to be returned.
Default is "item_name".}
}
\value{
A (named) list with string vectors of item names
}
\description{
Extracts a list of item names based on dic information
}
\details{
This function is basically a wrapper around the \code{select_items(data = data, filter = ..., names_only = TRUE)} function. It takes multiple filter
expressions for a single data frame and returns the item names. This is
mostly helpful for functions that take multiple scale definitions like
\code{alpha_table()}
}
\examples{
get_scales(ex_itrf,
  'APD' = subscale_2 == "APD",
  'OPP' = subscale_2 == "OPP",
  "SW" = subscale_2 == "SW",
  'AD' = subscale_2 == "AD"
)

}
