% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haven_dic.R
\name{haven_dic}
\alias{haven_dic}
\title{haven labels to dic files}
\usage{
haven_dic(data)
}
\arguments{
\item{data}{A data frame containing variables with haven labels}
}
\value{
A data frame with dic information
}
\description{
haven labels to dic files
}
