% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haven_dic.R
\name{haven_dic}
\alias{haven_dic}
\title{haven labels to dic files}
\usage{
haven_dic(data, remove_haven_class = FALSE)
}
\arguments{
\item{data}{A data frame containing variables with haven labels}

\item{remove_haven_class}{If TRUE, haven labels are removed.}
}
\value{
A data frame with dic information
}
\description{
haven labels to dic files
}
