% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_missing.R
\name{impute_missing}
\alias{impute_missing}
\title{Impute missing values}
\usage{
impute_missing(data, scale = NULL, subscale = NULL,
  subscale_2 = NULL)
}
\arguments{
\item{data}{A data frame}

\item{scale}{String with scale name}

\item{subscale}{String with subscale name}

\item{subscale_2}{String with sub subscale name}
}
\value{
A data frame with imputed data.
}
\description{
Impute missing values
}
