% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_missing.R
\name{impute_missing}
\alias{impute_missing}
\title{Impute missing values}
\usage{
impute_missing(
  data,
  scale = NULL,
  subscale = NULL,
  subscale_2 = NULL,
  force_to_scale = TRUE
)
}
\arguments{
\item{data}{A data frame}

\item{force_to_scale}{If TRUE, imputed values will be rounded and forced to the scale.
That is, a value below the scale's minimum or maximum will be set to the scale's minimum and maximum.}
}
\value{
A data frame with imputed data.
}
\description{
Impute missing values
}
