% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITRF.R, R/scaledic_data.R
\docType{data}
\name{ITRF}
\alias{ITRF}
\alias{dic_ITRF}
\title{Integrated Teacher Report Form}
\format{
A data frame with 4776 rows and 40 variables

An object of class \code{data.frame} with 4776 rows and 36 columns.
}
\usage{
ITRF

ITRF
}
\description{
A dataset containing the response teacher form data (40 items).

Data of the ITRF and a correponding dic file'
}
\keyword{datasets}
