% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITRF.R, R/scaledic_data.R
\docType{data}
\name{ITRF}
\alias{ITRF}
\alias{dic_ITRF}
\title{Integrated Teacher Report Formula for more than 5000 subjects}
\format{A data frame with 6643 rows and 40 variables}
\usage{
ITRF

ITRF
}
\description{
A dataset containing the response teacher formula data (40 items).

Data of the ITRF and a correponding dic file'
}
\keyword{datasets}
