% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan_model.R
\name{lavaan_model}
\alias{lavaan_model}
\title{Extract a model definition that can be applied to lavaan::cfa for a confirmatory factor analysis}
\usage{
lavaan_model(scales, adds = "", orthogonal = FALSE)
}
\arguments{
\item{scales}{A list with scale information}

\item{adds}{Additional model definitions}
}
\value{
A character string with a lavaan model definition
}
\description{
Extract a model definition that can be applied to lavaan::cfa for a confirmatory factor analysis
}
