% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_scales.R
\name{list_scales}
\alias{list_scales}
\title{List scales}
\usage{
list_scales(data, labels = FALSE)
}
\arguments{
\item{data}{The target data frame}

\item{labels}{If set TRUE, scale labels instead of abreviations are shown.}
}
\value{
A list with scales on diferent levels
}
\description{
List scales
}
