% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_scales.R
\name{list_scales}
\alias{list_scales}
\title{List scales}
\usage{
list_scales(data, labels = FALSE, n_items = FALSE)
}
\arguments{
\item{data}{The target data frame}

\item{labels}{If TRUE, scale labels instead of abreviations are shown.}

\item{n_items}{If TRUE, number of items for each scale, subscale, and sub_subscale is shown}
}
\value{
A list with scales on diferent levels
}
\description{
List scales
}
