% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_scales.R
\name{list_scales}
\alias{list_scales}
\title{List scales}
\usage{
list_scales(
  data,
  levels = c("scale", "subscale", "subscale_2"),
  n_items = FALSE,
  char_na = ""
)
}
\arguments{
\item{data}{The target data frame}

\item{levels}{Character vector with names of dic attributes used to extract scale information.}

\item{n_items}{If TRUE, number of items for each scale is shown}

\item{char_na}{Character for NAs.}
}
\value{
A data.frame with scales on different levels
}
\description{
List scales
}
