% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_norms.R
\name{lookup_norms}
\alias{lookup_norms}
\title{Look up norm table values}
\usage{
lookup_norms(
  rawscores,
  group = NULL,
  normtable,
  from = "raw",
  to = "T",
  group_label = names(group),
  label = NULL
)
}
\arguments{
\item{rawscores}{A vector with raw scores.}

\item{group}{A vector with group affiliations or a list with vectors for
multiple group categorizations.}

\item{normtable}{An excel file name or a data frame containing a norm table.}

\item{from}{Label of the raw score variable in file.}

\item{to}{Label of the norm score variable in file.}

\item{group_label}{Label of the group variable in file or a list with group
labels for multiple group categorizations.}

\item{label}{Item label of the resulting variable}
}
\value{
A vector with norm values.
}
\description{
Transforms raw values to norm values based on a norm table.
}
\examples{
normtable <- data.frame(
  age = rep(c(6, 8, 6, 8), each = 11),
  gender = rep(c("m", "w"), each = 22),
  raw =  rep(0:10, 4),
  T = rep(c(40, 42, 44, 46, 48, 50, 52, 54, 56, 58, 60,
            37, 39, 41, 43, 45, 47, 49, 51, 53, 55, 57), 2) + rep(c(0,5), each = 22)
)
rawscores <- c(5,5,3,1)
group_age <- c("6", "8", "6", "8")
group_gender <- c("m", "m", "w", "w")

lookup_norms(rawscores, group = list(age = group_age, gender = group_gender), normtable)

## When group values are not specified exactly, raw scores can be ambiguous:
lookup_norms(rawscores, group = list(gender = group_gender), normtable = normtable)

lookup_norms(rawscores, normtable = normtable)
}
