% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_norms.R
\name{lookup_norms}
\alias{lookup_norms}
\title{Look up norm table values}
\usage{
lookup_norms(
  rawscores,
  group,
  normtable,
  from = "raw",
  to = "T",
  group_label = "group"
)
}
\arguments{
\item{rawscores}{A vector with raw scores.}

\item{group}{A vector with group affiliations.}

\item{normtable}{An excel file name or a data frame containing a norm table.}

\item{from}{Label of the raw score variable in file.}

\item{to}{Label of the norm score variable in file.}

\item{group_label}{Label of the group variable in file.}
}
\value{
A vector with norm values.
}
\description{
Transforms raw values to norm values based on a norm table.
}
\examples{
normtable <- data.frame(
  group = c(6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8),
  raw = c(0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10),
  T = c(40, 42, 44, 46, 48, 50, 52, 54, 56, 58, 60, 37, 39, 41, 43, 45, 47,
        49, 51, 53, 55, 57)
)
rawscores <- c(5,5,3,1)
group <- c("6", "8", "6", "8")
lookup_norms(rawscores, group, normtable)

## When no group values are provided, raw scores are ambiguous:
lookup_norms(rawscores, normtable = normtable)
}
