% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names2.R
\name{names2item}
\alias{names2item}
\alias{names2label}
\title{Rename items
Set names of items as long labels and vice versa.}
\usage{
names2item(
  data,
  chars = NULL,
  prefix = "",
  char_sep = "_",
  char_weight = c("(-)", "(+)"),
  char_prefix_end = ": "
)

names2label(data)
}
\arguments{
\item{data}{A data frame}

\item{chars}{If not NULL, only the first n chars og the long label will be applied.}

\item{prefix}{A character string or vector of character strings defining a prefix.
May include "scale", "subscale", "subscale2", "index", "reverse", "weight", or "najme".}

\item{char_sep}{Character with seperator between prefix information.}

\item{char_weight}{Character vector of length two with signs for negative and positive weights.}

\item{char_prefix_end}{Character with seperator between prefi and item.}
}
\value{
A renamed data frame
}
\description{
Rename items
Set names of items as long labels and vice versa.
}
\details{
names2item renames to the long label.
names2label renames to the short label.
}
