% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names2.R
\name{names2item}
\alias{names2item}
\alias{names2label}
\title{Rename items
Set names of items as long labels and vice versa.}
\usage{
names2item(data, chars = NULL, prefix = "", short = FALSE,
  id = FALSE, reverse = FALSE, scale = NULL, sep = "_")

names2label(data)
}
\arguments{
\item{data}{A data frame}

\item{chars}{If not NULL, only the first n chars og the long label will be applied.}

\item{prefix}{A string defining a Prefix. Either "scale", "subscale", or "subscale2"}

\item{short}{If TRUE, the sublabel will be set before the variable name.}

\item{id}{If set TRUE, the item id will be set in the variable name}

\item{reverse}{If set TRUE, weights will be set in the varibable name}

\item{scale}{Unused}

\item{sep}{Character with seperator.}
}
\value{
A renamed data frame
}
\description{
Rename items
Set names of items as long labels and vice versa.
}
\details{
names2item renames to the long label.
names2label renames to the short label.
}
