% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dic-constructor.R
\name{new_dic}
\alias{new_dic}
\title{Low-level constructor for dic vectors}
\usage{
new_dic(x, dic_attributes = list(), label = NULL, labels = NULL)
}
\arguments{
\item{x}{An atomic vector (integer/double/logical/character/factor/Date, ...)}

\item{dic_attributes}{A named list with dic metadata.}

\item{label}{Optional haven-style variable label.}

\item{labels}{Optional haven-style value labels (named vector).}
}
\description{
Low-level constructor for dic vectors
}
\keyword{internal}
