% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dic-constructor.R
\name{new_dic}
\alias{new_dic}
\title{Dictionary class low level constructor}
\usage{
new_dic(
  x,
  item_name = NULL,
  item_label = NULL,
  scale = NULL,
  scale_label = NULL,
  values = NULL,
  value_labels = NULL,
  missing = NULL,
  weight = 1,
  type = "integer",
  class = "item",
  ...,
  .list = NULL
)
}
\arguments{
\item{x}{A variable}

\item{type}{defaults to "integer"}

\item{class}{default is "item"}

\item{...}{further dic arguments (e.g. `source = "James (1891)")}

\item{.list}{An alternative way to provide class arguments as a list
(overwrites previous arguments)}
}
\value{
An item of class dic.
}
\description{
Dictionary class low level constructor
}
\examples{
x <- new_dic(1:100, item_label = "The label of this item")
}
