% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dic-constructor.R
\name{new_dic}
\alias{new_dic}
\title{Dictionary class low level constructor}
\usage{
new_dic(
  x,
  item_name = NULL,
  item_label = NULL,
  values = NULL,
  value_labels = NULL,
  missing = NULL,
  weight = 1,
  type = NULL,
  recodes = NULL,
  class = "item",
  ...,
  .list = NULL,
  .coerce_class = TRUE,
  .message_attr = FALSE
)
}
\arguments{
\item{x}{A variable}

\item{item_name}{Character}

\item{item_label}{Character}

\item{values}{Numeric or character vector with values. The vector can be
named}

\item{value_labels}{Character of the form \verb{value = label; value2 = label2}}

\item{missing}{Numeric or character vector with values}

\item{weight}{numeric}

\item{type}{defaults to data type of x}

\item{recodes}{Recoding information e.g. \verb{4 = 1, .default = 0}}

\item{class}{default is "item"}

\item{...}{further dic arguments (e.g. \code{source = "James (1891)"})}

\item{.list}{An alternative way to provide class arguments as a list
(overwrites previous arguments)}

\item{.coerce_class}{Logical. If TRUE, tries to coerce classes of 'x' if
class does not match the \code{type} argument}

\item{.message_attr}{Logical. For internal use}
}
\value{
An item of class dic.
}
\description{
Dictionary class low level constructor
}
\examples{
x <- new_dic(1:100, item_label = "The label of this item")
}
