% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_dic.R
\name{new_dic}
\alias{new_dic}
\title{Compute a new scaledic variable}
\usage{
new_dic(fun, data = NULL, ...)
}
\arguments{
\item{fun}{A function}

\item{data}{A data frame}

\item{...}{dic parameters}
}
\description{
Compute a new scaledic variable
}
