% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_item.R
\name{rename_items}
\alias{rename_items}
\title{Rename items}
\usage{
rename_items(
  data,
  pattern = "{item_label}",
  max_chars = NULL,
  chars = max_chars
)
}
\arguments{
\item{data}{A data frame}

\item{pattern}{A character string with the syntax of the glue function (see
example).}

\item{max_chars, chars}{If not NULL, only the first n chars og the long label
will be applied.}
}
\value{
A renamed data frame
}
\description{
Rename items based on dic information.
}
\examples{
ex_itrf  |>
  rename_items(pattern = "{reverse}{name}: {label}")  |>
  names()

}
