% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_missing.R
\name{replace_missing}
\alias{replace_missing}
\title{Replace Missing values}
\usage{
replace_missing(data, replace = NA, report = FALSE)
}
\arguments{
\item{data}{A data frame}

\item{replace}{Replace value for missing values}

\item{report}{If TRUE, proportion of missing values is printed.}
}
\value{
A data frame with replaced missing values.
}
\description{
Replace Missing values
}
