% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_scaledic.R
\docType{package}
\name{scaledic-package}
\alias{scaledic-package}
\alias{scaledic}
\title{Scale Dictionary}
\description{
A collection of procedures to bring labels, scales, missing values etc. to a data frame.
}
\examples{
# apply a dictionary file to a data frame
dat <- apply_dic(ex_itrf, dic_ITRF)
# check for typos (not allowed values)
dat <- check_values(dat, replace = NA)
# Single imputation (EM algorith from the Amelia package)
# based on the variables of the provided scale
dat <- impute_missing(dat, scale == "ITRF" & subscale == "Ext")
dat <- impute_missing(dat, scale == "ITRF" & subscale == "Int")
# Show a table with all scales and scale labels included in the data frame
list_scales(dat, labels = TRUE)
# Example with pipeline syntax. Would be much easier to use the "describe" function
# from the psch packages instead of summarise_all here.
library(dplyr)
dat \%>\%
  select_items(scale == "ITRF" & subscale == "Ext") \%>\%
  rename_items(prefix = c("subscale2", "index")) \%>\%
  summarise_all(mean, na.rm = TRUE) \%>\%
  round(2) \%>\%
  t()
}
\author{
Juergen Wilbert [aut, cre]
}
\keyword{package}
