% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_from_dic.R
\name{score_from_dic}
\alias{score_from_dic}
\title{Automatic scale scoring from a dic file}
\usage{
score_from_dic(data, dic, impute_values = FALSE)
}
\arguments{
\item{data}{A data frame with dic information}

\item{dic}{A dic file with scoring information}

\item{impute_values}{If TRUE, missing values are imputed before scoring.}
}
\value{
A data frame with added scale scores
}
\description{
Automatic scale scoring from a dic file
}
\keyword{internal}
