% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_scales.R
\name{score_scale}
\alias{score_scale}
\title{Score scale
Calculates the scale scores.}
\usage{
score_scale(
  data,
  scale = NULL,
  subscale = NULL,
  subscale_2 = NULL,
  bind = FALSE,
  FUN = sum,
  ...
)
}
\arguments{
\item{data}{A data frame}

\item{bind}{If set TRUE, returns the complete data frame. If set false,
returns a data frame with the new score variables for each scale}

\item{FUN}{A function for caluclating the score (e.g., sum, mean, median)}

\item{...}{Further arguments passed to the FUN function (e.g., na.rm = TRUE)}
}
\value{
A data frame
}
\description{
Score scale
Calculates the scale scores.
}
