% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_scale.R
\name{score_scale}
\alias{score_scale}
\title{Score scale
Calculates the scale scores.}
\usage{
score_scale(
  data,
  filter,
  scale = NULL,
  subscale = NULL,
  subscale_2 = NULL,
  bind = FALSE,
  sum = FALSE,
  min_valid = 1,
  max_na = NA,
  label = NULL,
  FUN = NULL,
  ...
)
}
\arguments{
\item{data}{A data frame}

\item{filter}{A logical expression for any dic attribute (e.g. scale == "ITRF" & subscale == "Int")}

\item{bind}{If set TRUE, returns the complete data frame. If set false,
returns a data frame with the new score variables for each scale.}

\item{sum}{If TRUE, function mean(x, na.rm = TRUE) is applied for buiding the scores. If FALSE, function sum(x, na.rm = TRUE) is applied. When argument FUN is set, `sum` is ignored.}

\item{min_valid}{Minimal number of valid values that is required for calculating the mean.
A value between 0 and 1 indicates a proportion of values (e.g., 0.5 = 50 percent of values have to be valid).}

\item{max_na}{Maximum number of NAs that are allowed before returning NA.
A value between 0 and 1 indicates a proportion of values (e.g., 0.5 = 50 percent NAs are allowed).}

\item{label}{A character string with a label for the resulting score variable. Automatically generated if label is not set.}

\item{FUN}{A function for calculating the score (e.g., median)}

\item{...}{Further arguments passed to the FUN function (e.g., na.rm = TRUE)}
}
\value{
A data frame
}
\description{
Score scale
Calculates the scale scores.
}
