% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_items.R
\name{select_items}
\alias{select_items}
\title{Select items}
\usage{
select_items(data, filter = NULL, names_only = FALSE, index_only = FALSE)
}
\arguments{
\item{data}{A data frame with dic information}

\item{filter}{A logical expression for any dic attribute (e.g. scale == "ITRF" & subscale == "Int")}

\item{names_only}{If TRUE, variable names are returned instead of a data frame}

\item{index_only}{If TRUE, variable indices are returned instead of a data frame}
}
\value{
A data frame, a vector with variable names or a vector with indices
}
\description{
Selects a data frame with variables of a specific scale
}
