% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_scale.R
\name{select_scale}
\alias{select_scale}
\title{Select a scale
Selects a data frame with variables of a specific scale}
\usage{
select_scale(data, scale = NULL, subscale = NULL, subscale_2 = NULL)
}
\arguments{
\item{data}{A data frame}

\item{scale}{String with scale name}

\item{subscale}{String with subscale name}

\item{subscale_2}{String with sub subscale name}
}
\value{
A data frame
}
\description{
Select a scale
Selects a data frame with variables of a specific scale
}
