% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_scores.R
\name{select_scores}
\alias{select_scores}
\title{Select scale scores}
\usage{
select_scores(data, names_only = FALSE, index_only = FALSE)
}
\arguments{
\item{data}{A data frame}

\item{names_only}{If TRUE, variable names are returned instead of a data
frame}

\item{index_only}{If TRUE, variable indices are returned instead of a data
frame}
}
\value{
A data frame, a vector with variable names, or a vector with indices.
}
\description{
Selects variables generated by \code{\link[=score_scale]{score_scale()}} from a data frame.
}
