% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_dic.R
\name{set_dic}
\alias{set_dic}
\title{Set dictionary informations to variable}
\usage{
set_dic(data, vars, parameters, fill_missing = TRUE)
}
\arguments{
\item{data}{A data frame}

\item{vars}{string or vector with string of variable names}

\item{parameters}{List containing parameters}
}
\value{
A data frame
}
\description{
Set dictionary informations to variable
}
