% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_dic.R
\name{set_dic}
\alias{set_dic}
\title{Set dictionary information to variables}
\usage{
set_dic(data, .vars = NULL, ...)
}
\arguments{
\item{data}{A data frame or a vector.}

\item{.vars}{Character vector with variable names. If data is a data frame,
address these variables. If left \code{NULL} and data is a data frame, all
variables from the data frame are addressed.}

\item{...}{dic attributes of the form \code{attribute = value}.}
}
\value{
A data frame or a vector with added dic attributes.
}
\description{
Set dictionary information to variables
}
