% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dic.R
\name{[.dic}
\alias{[.dic}
\title{Subset a dic variable}
\usage{
\method{[}{dic}(x, i = length(x), ...)
}
\arguments{
\item{x}{An  object}

\item{i}{Elements to be selected}

\item{...}{further arguments passed to the subset function.}
}
\value{
A dic variable
}
\description{
Subset a dic variable
}
\keyword{internal}
