% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dic.R
\name{[.dic}
\alias{[.dic}
\title{Subset a dic variable}
\usage{
\method{[}{dic}(x, i = length(x), ...)
}
\arguments{
\item{x}{A scdf object}

\item{i}{A case name from x}
}
\value{
A dic variable
}
\description{
Subset a dic variable
}
