% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private.R
\name{which_dic}
\alias{which_dic}
\title{returns the index numbers of items in a data frame that contain dic values}
\usage{
which_dic(data, items_only = TRUE)
}
\description{
\itemize{
\item If item_only is FALSE then only dic infos for items of class item are returned.
}
}
\keyword{internal}
